EES-5097 Separate out DuckDb repository classes
