+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/ProcessorHostBuilder.cs
@@ -115,11 +115,16 @@ public static class ProcessorHostBuilder
                     .AddScoped<IDataSetService, DataSetService>()
                     .AddScoped<IDataSetMetaService, DataSetMetaService>()
                     .AddScoped<IDataSetVersionPathResolver, DataSetVersionPathResolver>()
-                    .AddScoped<IDataRepository, DataRepository>()
-                    .AddScoped<IFilterRepository, FilterRepository>()
-                    .AddScoped<IIndicatorRepository, IndicatorRepository>()
-                    .AddScoped<ILocationRepository, LocationRepository>()
-                    .AddScoped<ITimePeriodRepository, TimePeriodRepository>()
+                    .AddScoped<IDataDuckDbRepository, DataDuckDbRepository>()
+                    .AddScoped<IFilterOptionsDuckDbRepository, FilterOptionsDuckDbRepository>()
+                    .AddScoped<IIndicatorsDuckDbRepository, IndicatorsDuckDbRepository>()
+                    .AddScoped<ILocationsDuckDbRepository, LocationsDuckDbRepository>()
+                    .AddScoped<ITimePeriodsDuckDbRepository, TimePeriodsDuckDbRepository>()
+                    .AddScoped<IFilterMetaRepository, FilterMetaRepository>()
+                    .AddScoped<IGeographicLevelMetaRepository, GeographicLevelMetaRepository>()
+                    .AddScoped<IIndicatorMetaRepository, IndicatorMetaRepository>()
+                    .AddScoped<ILocationMetaRepository, LocationMetaRepository>()
+                    .AddScoped<ITimePeriodMetaRepository, TimePeriodMetaRepository>()
                     .AddScoped<IParquetService, ParquetService>()
                     .AddScoped<IPrivateBlobStorageService, PrivateBlobStorageService>()
                     .AddScoped<IValidator<InitialDataSetVersionCreateRequest>,
