+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/FilterMetaRepository.cs
@@ -1,7 +1,6 @@
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Database;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
-using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.Parquet.Tables;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Models;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Services.Interfaces;
@@ -13,9 +12,9 @@ using Microsoft.EntityFrameworkCore;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository;
 
-public class FilterRepository(
+public class FilterMetaRepository(
     PublicDataDbContext publicDataDbContext,
-    IDataSetVersionPathResolver dataSetVersionPathResolver) : IFilterRepository
+    IDataSetVersionPathResolver dataSetVersionPathResolver) : IFilterMetaRepository
 {
     public async Task CreateFilterMetas(
         IDuckDbConnection duckDbConnection,
@@ -126,49 +125,6 @@ public class FilterRepository(
                 throw new InvalidOperationException(
                     $"Inserted incorrect number of filter option meta links for {meta.PublicId}. " +
                     $"Inserted: {insertedLinks}, expected: {options.Count}");
-            }
-        }
-    }
-
-    public async Task CreateFilterMetaTable(
-        IDuckDbConnection duckDbConnection,
-        DataSetVersion dataSetVersion,
-        CancellationToken cancellationToken = default)
-    {
-        await publicDataDbContext
-            .Entry(dataSetVersion)
-            .Collection(dsv => dsv.FilterMetas)
-            .Query()
-            .Include(m => m.Options)
-            .LoadAsync(cancellationToken);
-
-        await duckDbConnection.SqlBuilder(
-            $"""
-             CREATE TABLE {FilterOptionsTable.TableName:raw}(
-                 {FilterOptionsTable.Cols.Id:raw} INTEGER PRIMARY KEY,
-                 {FilterOptionsTable.Cols.Label:raw} VARCHAR,
-                 {FilterOptionsTable.Cols.PublicId:raw} VARCHAR,
-                 {FilterOptionsTable.Cols.FilterId:raw} VARCHAR
-             )
-             """
-        ).ExecuteAsync(cancellationToken: cancellationToken);
-
-        var id = 1;
-
-        foreach (var filter in dataSetVersion.FilterMetas)
-        {
-            using var appender = duckDbConnection.CreateAppender(table: FilterOptionsTable.TableName);
-
-            foreach (var link in filter.OptionLinks.OrderBy(l => l.Option.Label))
-            {
-                var insertRow = appender.CreateRow();
-
-                insertRow.AppendValue(id++);
-                insertRow.AppendValue(link.Option.Label);
-                insertRow.AppendValue(link.PublicId);
-                insertRow.AppendValue(filter.PublicId);
-
-                insertRow.EndRow();
             }
         }
     }
