+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/IDataDuckDbRepository.cs
@@ -1,0 +1,10 @@
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
+
+public interface IDataDuckDbRepository
+{
+    Task CreateDataTable(
+        DataSetVersion dataSetVersion,
+        CancellationToken cancellationToken = default);
+}
