+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/IFilterMetaRepository.cs
@@ -1,10 +1,13 @@
 using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
 
-public interface IDataRepository
+public interface IFilterMetaRepository
 {
-    Task CreateDataTable(
+    Task CreateFilterMetas(
+        IDuckDbConnection duckDbConnection,
         DataSetVersion dataSetVersion,
+        IReadOnlySet<string> allowedColumns,
         CancellationToken cancellationToken = default);
 }
