+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/IGeographicLevelMetaRepository.cs
@@ -1,0 +1,12 @@
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
+
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
+
+public interface IGeographicLevelMetaRepository
+{
+    Task<GeographicLevelMeta> CreateGeographicLevelMeta(
+        IDuckDbConnection duckDb,
+        DataSetVersion dataSetVersion,
+        CancellationToken cancellationToken = default);
+}
