+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/IIndicatorMetaRepository.cs
@@ -3,21 +3,11 @@ using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
 
-public interface ILocationRepository
+public interface IIndicatorMetaRepository
 {
-    Task<GeographicLevelMeta> CreateGeographicLevelMeta(
-        IDuckDbConnection duckDb,
-        DataSetVersion dataSetVersion,
-        CancellationToken cancellationToken = default);
-
-    Task CreateLocationMetas(
+    Task CreateIndicatorMetas(
         IDuckDbConnection duckDbConnection,
         DataSetVersion dataSetVersion,
         IReadOnlySet<string> allowedColumns,
-        CancellationToken cancellationToken = default);
-
-    Task CreateLocationMetaTable(
-        IDuckDbConnection duckDbConnection,
-        DataSetVersion dataSetVersion,
         CancellationToken cancellationToken = default);
 }
