+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/ILocationMetaRepository.cs
@@ -3,16 +3,11 @@ using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
 
 namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
 
-public interface IFilterRepository
+public interface ILocationMetaRepository
 {
-    Task CreateFilterMetas(
+    Task CreateLocationMetas(
         IDuckDbConnection duckDbConnection,
         DataSetVersion dataSetVersion,
         IReadOnlySet<string> allowedColumns,
-        CancellationToken cancellationToken = default);
-
-    Task CreateFilterMetaTable(
-        IDuckDbConnection duckDbConnection,
-        DataSetVersion dataSetVersion,
         CancellationToken cancellationToken = default);
 }
