+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/ILocationsDuckDbRepository.cs
@@ -1,0 +1,12 @@
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
+
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
+
+public interface ILocationsDuckDbRepository
+{
+    Task CreateLocationsTable(
+        IDuckDbConnection duckDbConnection,
+        DataSetVersion dataSetVersion,
+        CancellationToken cancellationToken = default);
+}
