+++ src/GovUk.Education.ExploreEducationStatistics.Public.Data.Processor/Repository/Interfaces/ITimePeriodsDuckDbRepository.cs
@@ -1,0 +1,12 @@
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model;
+using GovUk.Education.ExploreEducationStatistics.Public.Data.Model.DuckDb;
+
+namespace GovUk.Education.ExploreEducationStatistics.Public.Data.Processor.Repository.Interfaces;
+
+public interface ITimePeriodsDuckDbRepository
+{
+    Task CreateTimePeriodsTable(
+        IDuckDbConnection duckDbConnection,
+        DataSetVersion dataSetVersion,
+        CancellationToken cancellationToken = default);
+}
